requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
joblib
mysql-connector-python
